Or this
Note that the regex patterns used in this script assume that the HTML file is well-formed and consistent. If the file structure changes or becomes inconsistent, the patterns may need to be updated accordingly. Additionally, this script assumes that there are no ties for the mode; if there are ties, the script will only return one of the most common values.